+++ src/components/forms/set-password-form.tsx
@@ -6,7 +6,7 @@ import styled from "styled-components";
 import { FormErrorMessage } from "../input/form-error-message";
 import { InputText } from "../input/input-text";
 import { StyledForm } from "./form";
-import { Button, ButtonVariant } from "@src/components/buttons/buttons";
+import { Button, ButtonVariant } from "@src/components/buttons";
 import { PasswordRulesErrorList } from "@src/components/password-rules-error-list";
 import { formatErrorMessage } from "@src/configs/intl";
 import { ERRORS_DATA } from "@src/errors/web-errors";
