+++ tests/unit/form.test.tsx
@@ -1,7 +1,7 @@
 import userEvent from "@testing-library/user-event";
 import React from "react";
 
-import { Button, ButtonVariant } from "../../src/components/buttons/buttons";
+import { Button, ButtonVariant } from "../../src/components/buttons";
 import { Form } from "../../src/components/forms/form";
 import { render, screen } from "../config/testing-library-config";
